requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
docopt==0.6.2
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
GeoAlchemy2==0.4.2
gevent==1.2.2
greenlet==0.4.13
idna==2.6
itsdangerous==0.24
Jinja2==2.10.1
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.3.0
pipreqs==0.4.9
pluggy==0.7.1
pstat==0.0.1
psutil==5.4.3
psycogreen==1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.6.0
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.6.4
pyOpenSSL==17.5.0
pytest==3.7.4
python-dateutil==2.7.3
python-editor==1.0.3
requests==2.21.0
six==1.11.0
SQLAlchemy==1.3.3
urllib3==1.24.2
Werkzeug==0.14.1
yarg==0.1.9
